requirements.txt
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cssselect==1.4.0
et_xmlfile==2.0.0
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.25.2
Flask==3.1.3
idna==3.11
itsdangerous==2.2.0
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.5.3
kagglehub==1.0.0
kagglesdk==0.1.16
lxml==6.0.2
lxml_html_clean==0.4.4
MarkupSafe==3.0.3
newspaper3k==0.2.8
nltk==3.9.4
numpy==2.4.4
openpyxl==3.1.5
packaging==26.0
pandas==3.0.2
pillow==12.2.0
protobuf==7.34.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
requests-file==3.0.1
scikit-learn==1.8.0
scipy==1.17.1
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.8.3
threadpoolctl==3.6.0
tinysegmenter==0.3
tldextract==5.3.1
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
Werkzeug==3.1.8
